Move camelCase test for telegram.ext.ExtBot into test_bot and add telegram.Bot camelCase test
